+++ src/Coercer/Boolean.php
@@ -6,7 +6,6 @@ namespace Duon\Sire\Coercer;
 
 use Duon\Sire\Coercion;
 use Duon\Sire\Contract;
-use Duon\Sire\Value;
 use Override;
 
 final readonly class Boolean implements Contract\Coercer
@@ -20,25 +19,26 @@ final readonly class Boolean implements Contract\Coercer
 	public function coerce(mixed $pristine, string $label): Contract\Coercion
 	{
 		if (is_bool($pristine)) {
-			return new Coercion(new Value($pristine, $pristine));
+			return new Coercion($pristine, $pristine);
 		}
 
 		if (!$pristine) {
-			return new Coercion(new Value(false, $pristine));
+			return new Coercion(false, $pristine);
 		}
 
 		$tmp = strtolower((string) $pristine);
 
 		if (in_array($tmp, ['1', 'on', 'true', 'yes'], true)) {
-			return new Coercion(new Value(true, $pristine));
+			return new Coercion(true, $pristine);
 		}
 
 		if (in_array($tmp, ['0', 'off', 'false', 'no', 'null'], true)) {
-			return new Coercion(new Value(false, $pristine));
+			return new Coercion(false, $pristine);
 		}
 
 		return new Coercion(
-			new Value($pristine, $pristine),
+			$pristine,
+			$pristine,
 			sprintf($this->messages['bool'], $label),
 		);
 	}