+++ src/Coercer/FloatingPoint.php
@@ -6,7 +6,6 @@ namespace Duon\Sire\Coercer;
 
 use Duon\Sire\Coercion;
 use Duon\Sire\Contract;
-use Duon\Sire\Value;
 use Override;
 
 final readonly class FloatingPoint implements Contract\Coercer
@@ -20,21 +19,22 @@ final readonly class FloatingPoint implements Contract\Coercer
 	public function coerce(mixed $pristine, string $label): Contract\Coercion
 	{
 		if (is_float($pristine) || is_null($pristine)) {
-			return new Coercion(new Value($pristine, $pristine));
+			return new Coercion($pristine, $pristine);
 		}
 
 		if (is_int($pristine)) {
-			return new Coercion(new Value((float) $pristine, $pristine));
+			return new Coercion((float) $pristine, $pristine);
 		}
 
 		$tmp = trim((string) $pristine);
 
 		if (preg_match('/^[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?$/', $tmp)) {
-			return new Coercion(new Value((float) $tmp, $pristine));
+			return new Coercion((float) $tmp, $pristine);
 		}
 
 		return new Coercion(
-			new Value($pristine, $pristine),
+			$pristine,
+			$pristine,
 			sprintf($this->messages['float'], $label),
 		);
 	}