+++ src/Coercer/Integer.php
@@ -6,7 +6,6 @@ namespace Duon\Sire\Coercer;
 
 use Duon\Sire\Coercion;
 use Duon\Sire\Contract;
-use Duon\Sire\Value;
 use Override;
 
 final readonly class Integer implements Contract\Coercer
@@ -20,15 +19,16 @@ final readonly class Integer implements Contract\Coercer
 	public function coerce(mixed $pristine, string $label): Contract\Coercion
 	{
 		if (is_int($pristine) || is_null($pristine)) {
-			return new Coercion(new Value($pristine, $pristine));
+			return new Coercion($pristine, $pristine);
 		}
 
 		if (preg_match('/^([0-9]|-[1-9]|-?[1-9][0-9]*)$/i', trim((string) $pristine))) {
-			return new Coercion(new Value((int) $pristine, $pristine));
+			return new Coercion((int) $pristine, $pristine);
 		}
 
 		return new Coercion(
-			new Value($pristine, $pristine),
+			$pristine,
+			$pristine,
 			sprintf($this->messages['int'], $label),
 		);
 	}