+++ src/Coercer/Sequence.php
@@ -6,7 +6,6 @@ namespace Duon\Sire\Coercer;
 
 use Duon\Sire\Coercion;
 use Duon\Sire\Contract;
-use Duon\Sire\Value;
 use Override;
 
 final readonly class Sequence implements Contract\Coercer
@@ -23,11 +22,12 @@ final readonly class Sequence implements Contract\Coercer
 			is_array($pristine)
 			&& ($pristine === [] || array_keys($pristine) === range(0, count($pristine) - 1))
 		) {
-			return new Coercion(new Value($pristine, $pristine));
+			return new Coercion($pristine, $pristine);
 		}
 
 		return new Coercion(
-			new Value($pristine, $pristine),
+			$pristine,
+			$pristine,
 			sprintf($this->messages['list'], $label),
 		);
 	}