+++ tests/CoercerRegistryTest.php
@@ -9,7 +9,6 @@ use Duon\Sire\CoercerRegistry;
 use Duon\Sire\Coercion;
 use Duon\Sire\Contract;
 use Duon\Sire\Contract\Coercer;
-use Duon\Sire\Value;
 use Override;
 use RuntimeException;
 
@@ -101,7 +100,7 @@ class CoercerRegistryTest extends TestCase
 			#[Override]
 			public function coerce(mixed $pristine, string $label): Contract\Coercion
 			{
-				return new Coercion(new Value(($this->callback)($pristine), $pristine));
+				return new Coercion(($this->callback)($pristine), $pristine);
 			}
 		};
 	}