+++ tests/CoercionTest.php
@@ -5,22 +5,21 @@ declare(strict_types=1);
 namespace Duon\Sire\Tests;
 
 use Duon\Sire\Coercion;
-use Duon\Sire\Value;
 
 class CoercionTest extends TestCase
 {
 	public function testProperties(): void
 	{
-		$value = new Value('coerced', 'raw');
-		$coercion = new Coercion($value, 'Invalid');
+		$coercion = new Coercion('coerced', 'raw', 'Invalid');
 
-		$this->assertSame($value, $coercion->value);
+		$this->assertSame('coerced', $coercion->value);
+		$this->assertSame('raw', $coercion->pristine);
 		$this->assertSame('Invalid', $coercion->error);
 	}
 
 	public function testErrorDefaultsToNull(): void
 	{
-		$coercion = new Coercion(new Value('coerced', 'raw'));
+		$coercion = new Coercion('coerced', 'raw');
 
 		$this->assertNull($coercion->error);
 	}